Change size of hopCount: uint8_t -> uint16_t.

+++ MAC/include/ControlPacket.h
@@ -28,7 +28,7 @@ typedef struct
     uint8_t     currentSlot;
     uint8_t     collisionSlot;
     uint32_t    collisionFrequency;
-    uint8_t     hopCount;
+    uint16_t     hopCount;
     uint64_t    networkTime;
     uint32_t    initTime;
     uint8_t     ackSize;
@@ -37,7 +37,7 @@ typedef struct
 
 void controlpacket_init(ControlPacket_t DOUBLE_POINTER pkt);
 void controlpacket_destroy(ControlPacket_t DOUBLE_POINTER pkt);
-void controlpacket_create(ControlPacket_t *pkt, uint16_t nodeID, uint32_t frame, uint8_t slot, uint8_t collisionSlots, uint32_t collisionFrequency, uint8_t hopCount, uint64_t netTime, uint32_t initTime, ARRAY *ack, int n);
+void controlpacket_create(ControlPacket_t *pkt, uint16_t nodeID, uint32_t frame, uint8_t slot, uint8_t collisionSlots, uint32_t collisionFrequency, uint16_t hopCount, uint64_t netTime, uint32_t initTime, ARRAY *ack, int n);
 void controlpacket_clear(ControlPacket_t *pkt);
 void controlpacket_set_nodeID(ControlPacket_t *pkt, uint16_t nodeID);
 uint16_t controlpacket_get_nodeID(ControlPacket_t *pkt);
@@ -49,8 +49,8 @@ void controlpacket_set_collision_slot(ControlPacket_t *pkt, uint8_t slot);
 uint8_t controlpacket_get_collision_slot(ControlPacket_t *pkt);
 void controlpacket_set_collision_frequency(ControlPacket_t *pkt, uint32_t frequency);
 uint32_t controlpacket_get_collision_frequency(ControlPacket_t *pkt);
-void controlpacket_set_hop_count(ControlPacket_t *pkt, uint8_t count);
-uint8_t controlpacket_get_hop_count(ControlPacket_t *pkt);
+void controlpacket_set_hop_count(ControlPacket_t *pkt, uint16_t count);
+uint16_t controlpacket_get_hop_count(ControlPacket_t *pkt);
 void controlpacket_set_network_time(ControlPacket_t *pkt, uint64_t time);
 uint64_t controlpacket_get_network_time(ControlPacket_t *pkt);
 void controlpacket_set_init_time(ControlPacket_t *pkt, uint32_t time);
